backend/requirements.txt
fastapi==0.103.2
uvicorn==0.24.0
python-dotenv==1.0.0
python-multipart==0.0.6
pydantic==2.5.0
httpx==0.24.1
supabase==1.2.0
passlib[bcrypt]==1.7.4
